moved label_images method to utils
